# apps/api/openapi.json
{"components":{"schemas":{"AccordDetail":{"properties":{"fragrances":{"$ref":"#/components/schemas/ListEnvelope_FragranceListItem_"},"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"}},"required":["slug","name","fragrances"],"title":"AccordDetail","type":"object"},"AccordSummary":{"properties":{"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"}},"required":["slug","name"],"title":"AccordSummary","type":"object"},"AccordsListResponse":{"properties":{"data":{"items":{"$ref":"#/components/schemas/AccordSummary"},"title":"Data","type":"array"}},"required":["data"],"title":"AccordsListResponse","type":"object"},"ArticleDetail":{"properties":{"body":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Body"},"fragrances":{"items":{"$ref":"#/components/schemas/FragranceListItem"},"title":"Fragrances","type":"array"},"published_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Published At"},"slug":{"title":"Slug","type":"string"},"title":{"title":"Title","type":"string"}},"required":["slug","title","body","published_at","fragrances"],"title":"ArticleDetail","type":"object"},"ArticleSummary":{"properties":{"published_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Published At"},"slug":{"title":"Slug","type":"string"},"title":{"title":"Title","type":"string"}},"required":["slug","title"],"title":"ArticleSummary","type":"object"},"BrandDetail":{"properties":{"fragrances":{"$ref":"#/components/schemas/ListEnvelope_FragranceListItem_"},"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"}},"required":["slug","name","fragrances"],"title":"BrandDetail","type":"object"},"BrandSummary":{"properties":{"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"}},"required":["slug","name"],"title":"BrandSummary","type":"object"},"ConcentrationSummary":{"properties":{"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"}},"required":["slug","name"],"title":"ConcentrationSummary","type":"object"},"FilterSpec":{"additionalProperties":false,"description":"Structured catalog filters; mirrors the P1 list-endpoint filters.","properties":{"accord":{"items":{"type":"string"},"title":"Accord","type":"array"},"brand":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Brand"},"concentration":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Concentration"},"gender":{"anyOf":[{"enum":["masculine","feminine","unisex"],"type":"string"},{"type":"null"}],"title":"Gender"},"note":{"items":{"type":"string"},"title":"Note","type":"array"},"year_max":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year Max"},"year_min":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year Min"}},"title":"FilterSpec","type":"object"},"FragranceDetail":{"properties":{"accords":{"items":{"$ref":"#/components/schemas/AccordSummary"},"title":"Accords","type":"array"},"articles":{"items":{"$ref":"#/components/schemas/ArticleSummary"},"title":"Articles","type":"array"},"brand":{"$ref":"#/components/schemas/BrandSummary"},"concentration":{"anyOf":[{"$ref":"#/components/schemas/ConcentrationSummary"},{"type":"null"}]},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"gender":{"title":"Gender","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"notes":{"$ref":"#/components/schemas/NotesByRole"},"perfumers":{"items":{"$ref":"#/components/schemas/PerfumerSummary"},"title":"Perfumers","type":"array"},"slug":{"title":"Slug","type":"string"},"year_released":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year Released"},"year_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Year Text"}},"required":["id","slug","name","description","year_released","year_text","gender","brand","concentration","perfumers","notes","accords","articles"],"title":"FragranceDetail","type":"object"},"FragranceListItem":{"properties":{"brand":{"$ref":"#/components/schemas/BrandSummary"},"concentration":{"anyOf":[{"$ref":"#/components/schemas/ConcentrationSummary"},{"type":"null"}]},"gender":{"title":"Gender","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"},"year_released":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year Released"}},"required":["id","slug","name","brand","year_released","gender","concentration"],"title":"FragranceListItem","type":"object"},"Gender":{"description":"Maps to PG ENUM ``fragrance_gender``.","enum":["masc","fem","unisex","genderfree"],"title":"Gender","type":"string"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"ListEnvelope_ArticleSummary_":{"properties":{"data":{"items":{"$ref":"#/components/schemas/ArticleSummary"},"title":"Data","type":"array"},"pagination":{"$ref":"#/components/schemas/Pagination"}},"required":["data","pagination"],"title":"ListEnvelope[ArticleSummary]","type":"object"},"ListEnvelope_BrandSummary_":{"properties":{"data":{"items":{"$ref":"#/components/schemas/BrandSummary"},"title":"Data","type":"array"},"pagination":{"$ref":"#/components/schemas/Pagination"}},"required":["data","pagination"],"title":"ListEnvelope[BrandSummary]","type":"object"},"ListEnvelope_FragranceListItem_":{"properties":{"data":{"items":{"$ref":"#/components/schemas/FragranceListItem"},"title":"Data","type":"array"},"pagination":{"$ref":"#/components/schemas/Pagination"}},"required":["data","pagination"],"title":"ListEnvelope[FragranceListItem]","type":"object"},"ListEnvelope_PerfumerSummary_":{"properties":{"data":{"items":{"$ref":"#/components/schemas/PerfumerSummary"},"title":"Data","type":"array"},"pagination":{"$ref":"#/components/schemas/Pagination"}},"required":["data","pagination"],"title":"ListEnvelope[PerfumerSummary]","type":"object"},"NoteDetail":{"properties":{"fragrances":{"$ref":"#/components/schemas/ListEnvelope_FragranceListItem_"},"name":{"title":"Name","type":"string"},"parent":{"anyOf":[{"$ref":"#/components/schemas/NoteSummary"},{"type":"null"}]},"slug":{"title":"Slug","type":"string"}},"required":["slug","name","parent","fragrances"],"title":"NoteDetail","type":"object"},"NoteSummary":{"properties":{"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"}},"required":["slug","name"],"title":"NoteSummary","type":"object"},"NoteTreeEnvelope":{"properties":{"data":{"items":{"$ref":"#/components/schemas/NoteTreeNode"},"title":"Data","type":"array"}},"required":["data"],"title":"NoteTreeEnvelope","type":"object"},"NoteTreeNode":{"description":"Self-recursive tree node used by GET /api/v1/notes.","properties":{"children":{"items":{"$ref":"#/components/schemas/NoteTreeNode"},"title":"Children","type":"array"},"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"}},"required":["slug","name"],"title":"NoteTreeNode","type":"object"},"NotesByRole":{"properties":{"base":{"default":[],"items":{"$ref":"#/components/schemas/NoteSummary"},"title":"Base","type":"array"},"heart":{"default":[],"items":{"$ref":"#/components/schemas/NoteSummary"},"title":"Heart","type":"array"},"top":{"default":[],"items":{"$ref":"#/components/schemas/NoteSummary"},"title":"Top","type":"array"}},"title":"NotesByRole","type":"object"},"Pagination":{"properties":{"has_next":{"title":"Has Next","type":"boolean"},"limit":{"title":"Limit","type":"integer"},"offset":{"title":"Offset","type":"integer"},"total":{"title":"Total","type":"integer"}},"required":["limit","offset","total","has_next"],"title":"Pagination","type":"object"},"PerfumerDetail":{"properties":{"fragrances":{"$ref":"#/components/schemas/ListEnvelope_FragranceListItem_"},"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"}},"required":["slug","name","fragrances"],"title":"PerfumerDetail","type":"object"},"PerfumerSummary":{"properties":{"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"}},"required":["slug","name"],"title":"PerfumerSummary","type":"object"},"SearchHit":{"description":"One ranked candidate in the search response.","properties":{"fragrance":{"$ref":"#/components/schemas/FragranceListItem"},"match_reason":{"items":{"enum":["vector","fts","ontology"],"type":"string"},"title":"Match Reason","type":"array"},"relevance_score":{"maximum":1.0,"minimum":0.0,"title":"Relevance Score","type":"number"}},"required":["fragrance","relevance_score","match_reason"],"title":"SearchHit","type":"object"},"SearchRequest":{"additionalProperties":false,"description":"`POST /api/v1/search` body.","properties":{"filters":{"$ref":"#/components/schemas/FilterSpec"},"include":{"items":{"enum":["notes","brand","perfumer"],"type":"string"},"title":"Include","type":"array"},"query":{"maxLength":512,"minLength":1,"title":"Query","type":"string"},"top_k":{"default":20,"maximum":50.0,"minimum":1.0,"title":"Top K","type":"integer"}},"required":["query"],"title":"SearchRequest","type":"object"},"SearchResponse":{"description":"Envelope returned by both `/search` and `/{slug}/similar`.","properties":{"cache_hit":{"default":false,"title":"Cache Hit","type":"boolean"},"data":{"items":{"$ref":"#/components/schemas/SearchHit"},"title":"Data","type":"array"},"degraded":{"default":false,"title":"Degraded","type":"boolean"}},"required":["data"],"title":"SearchResponse","type":"object"},"SimilarRequest":{"additionalProperties":false,"description":"`POST /api/v1/fragrances/{slug}/similar` body.","properties":{"include":{"items":{"enum":["notes","brand","perfumer"],"type":"string"},"title":"Include","type":"array"},"top_k":{"default":20,"maximum":50.0,"minimum":1.0,"title":"Top K","type":"integer"}},"title":"SimilarRequest","type":"object"},"ValidationError":{"properties":{"ctx":{"title":"Context","type":"object"},"input":{"title":"Input"},"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"}}},"info":{"title":"Fragwise API","version":"0.1.0"},"openapi":"3.1.0","paths":{"/api/v1/accords":{"get":{"operationId":"list_accords_api_v1_accords_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccordsListResponse"}}},"description":"Successful Response"}},"summary":"List Accords","tags":["Accords"]}},"/api/v1/accords/{slug}":{"get":{"operationId":"get_accord_api_v1_accords__slug__get","parameters":[{"in":"path","name":"slug","required":true,"schema":{"pattern":"^[a-z0-9-]+$","title":"Slug","type":"string"}},{"in":"query","name":"limit","required":false,"schema":{"default":20,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}},{"in":"query","name":"offset","required":false,"schema":{"default":0,"minimum":0,"title":"Offset","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccordDetail"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Accord","tags":["Accords"]}},"/api/v1/articles":{"get":{"operationId":"list_articles_api_v1_articles_get","parameters":[{"in":"query","name":"limit","required":false,"schema":{"default":20,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}},{"in":"query","name":"offset","required":false,"schema":{"default":0,"minimum":0,"title":"Offset","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListEnvelope_ArticleSummary_"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Articles","tags":["Articles"]}},"/api/v1/articles/{slug}":{"get":{"operationId":"get_article_api_v1_articles__slug__get","parameters":[{"in":"path","name":"slug","required":true,"schema":{"pattern":"^[a-z0-9-]+$","title":"Slug","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleDetail"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Article","tags":["Articles"]}},"/api/v1/brands":{"get":{"operationId":"list_brands_api_v1_brands_get","parameters":[{"in":"query","name":"limit","required":false,"schema":{"default":20,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}},{"in":"query","name":"offset","required":false,"schema":{"default":0,"minimum":0,"title":"Offset","type":"integer"}},{"in":"query","name":"q","required":false,"schema":{"anyOf":[{"maxLength":100,"type":"string"},{"type":"null"}],"title":"Q"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListEnvelope_BrandSummary_"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Brands","tags":["Brands"]}},"/api/v1/brands/{slug}":{"get":{"operationId":"get_brand_api_v1_brands__slug__get","parameters":[{"in":"path","name":"slug","required":true,"schema":{"pattern":"^[a-z0-9-]+$","title":"Slug","type":"string"}},{"in":"query","name":"limit","required":false,"schema":{"default":20,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}},{"in":"query","name":"offset","required":false,"schema":{"default":0,"minimum":0,"title":"Offset","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrandDetail"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Brand","tags":["Brands"]}},"/api/v1/fragrances":{"get":{"operationId":"list_fragrances_api_v1_fragrances_get","parameters":[{"in":"query","name":"limit","required":false,"schema":{"default":20,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}},{"in":"query","name":"offset","required":false,"schema":{"default":0,"minimum":0,"title":"Offset","type":"integer"}},{"in":"query","name":"brand","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Brand"}},{"in":"query","name":"perfumer","required":false,"schema":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Perfumer"}},{"in":"query","name":"gender","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/Gender"},{"type":"null"}],"title":"Gender"}},{"in":"query","name":"year_min","required":false,"schema":{"anyOf":[{"maximum":2100,"minimum":1700,"type":"integer"},{"type":"null"}],"title":"Year Min"}},{"in":"query","name":"year_max","required":false,"schema":{"anyOf":[{"maximum":2100,"minimum":1700,"type":"integer"},{"type":"null"}],"title":"Year Max"}},{"in":"query","name":"concentration","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Concentration"}},{"in":"query","name":"accord","required":false,"schema":{"anyOf":[{"items":{"type":"string"},"maxItems":10,"type":"array"},{"type":"null"}],"title":"Accord"}},{"in":"query","name":"note","required":false,"schema":{"anyOf":[{"items":{"type":"string"},"maxItems":20,"type":"array"},{"type":"null"}],"title":"Note"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListEnvelope_FragranceListItem_"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Fragrances","tags":["Fragrances"]}},"/api/v1/fragrances/{slug}":{"get":{"operationId":"get_fragrance_api_v1_fragrances__slug__get","parameters":[{"in":"path","name":"slug","required":true,"schema":{"pattern":"^[a-z0-9-]+$","title":"Slug","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FragranceDetail"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Fragrance","tags":["Fragrances"]}},"/api/v1/fragrances/{slug}/similar":{"post":{"description":"Return fragrances similar to `{slug}` using its stored embedding.\n\nShares the per-IP rate limit and daily kill-switch with `/search`.\nDoes NOT call OpenAI: the embedding is loaded from `fragrance_embeddings`.","operationId":"similar_fragrances_api_v1_fragrances__slug__similar_post","parameters":[{"in":"path","name":"slug","required":true,"schema":{"pattern":"^[a-z0-9-]+$","title":"Slug","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SimilarRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Similar Fragrances","tags":["Fragrances"]}},"/api/v1/notes":{"get":{"operationId":"list_notes_api_v1_notes_get","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteTreeEnvelope"}}},"description":"Successful Response"}},"summary":"List Notes","tags":["Notes"]}},"/api/v1/notes/{slug}":{"get":{"operationId":"get_note_api_v1_notes__slug__get","parameters":[{"in":"path","name":"slug","required":true,"schema":{"pattern":"^[a-z0-9-]+$","title":"Slug","type":"string"}},{"in":"query","name":"limit","required":false,"schema":{"default":20,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}},{"in":"query","name":"offset","required":false,"schema":{"default":0,"minimum":0,"title":"Offset","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteDetail"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Note","tags":["Notes"]}},"/api/v1/perfumers":{"get":{"operationId":"list_perfumers_api_v1_perfumers_get","parameters":[{"in":"query","name":"limit","required":false,"schema":{"default":20,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}},{"in":"query","name":"offset","required":false,"schema":{"default":0,"minimum":0,"title":"Offset","type":"integer"}},{"in":"query","name":"q","required":false,"schema":{"anyOf":[{"maxLength":100,"type":"string"},{"type":"null"}],"title":"Q"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListEnvelope_PerfumerSummary_"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Perfumers","tags":["Perfumers"]}},"/api/v1/perfumers/{slug}":{"get":{"operationId":"get_perfumer_api_v1_perfumers__slug__get","parameters":[{"in":"path","name":"slug","required":true,"schema":{"pattern":"^[a-z0-9-]+$","title":"Slug","type":"string"}},{"in":"query","name":"limit","required":false,"schema":{"default":20,"maximum":100,"minimum":1,"title":"Limit","type":"integer"}},{"in":"query","name":"offset","required":false,"schema":{"default":0,"minimum":0,"title":"Offset","type":"integer"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PerfumerDetail"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Perfumer","tags":["Perfumers"]}},"/api/v1/search":{"post":{"description":"Hybrid pgvector + FTS + ontology search.","operationId":"search_api_v1_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Search","tags":["Search"]}},"/healthz":{"get":{"description":"Liveness probe. MUST NOT touch DB/Redis.","operationId":"healthz_healthz_get","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response Healthz Healthz Get","type":"object"}}},"description":"Successful Response"}},"summary":"Healthz"}},"/readyz":{"get":{"description":"Readiness probe. S7: error message is generic; full exception is\nlogged server-side so connection strings, hostnames, or driver-leaked\ncredentials never reach the response body.","operationId":"readyz_readyz_get","responses":{"200":{"content":{"application/json":{"schema":{"title":"Response Readyz Readyz Get"}}},"description":"Successful Response"}},"summary":"Readyz"}}}}